fix(api): change Supertrend multiplier type from float to int fix(twelvedata): change Supertrend multiplier type from float to int fix(frontend): update date formatting function for prefetch last run

+++ src/api/routes.py
@@ -4595,7 +4595,7 @@ async def get_supertrend(
     symbol: str,
     interval: str = "1day",
     period: int = 10,
-    multiplier: float = 3.0,
+    multiplier: int = 3,
     outputsize: int = 100,
 ):
     """Get Supertrend indicator."""
